Rename DisplayGroupService
